Moved homepage into app folder
Move the index/homepage to the app folder. Redirect both the index.android.js and index.ios.js to point to app/HomePage.js. Also change import routes for login, register, home, root at HomePage.js
